feat: ensure files are persisted to disk at every flush event if network is not accepting them, else send to network
