--- src/audit.rs
@@ -1,339 +1,0 @@
-/*
- * Parseable Server (C) 2022 - 2024 Parseable, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-use std::{
-    collections::{HashMap, VecDeque},
-    fmt::{Debug, Display},
-};
-
-use crate::{about::current, storage::StorageMetadata, HTTP_CLIENT};
-
-use super::option::CONFIG;
-use chrono::{DateTime, Utc};
-use once_cell::sync::Lazy;
-use serde::Serialize;
-use tracing::error;
-
-use tokio::{sync::Mutex, time::interval};
-use ulid::Ulid;
-use url::Url;
-
-// Shared audit logger instance to batch and send audit logs
-static AUDIT_LOGGER: Lazy<AuditLogger> = Lazy::new(AuditLogger::new);
-
-// AuditLogger handles sending audit logs to a remote logging system
-pub struct AuditLogger {
-    log_endpoint: Option<Url>,
-    batch: Mutex<VecDeque<AuditLog>>,
-}
-
-impl AuditLogger {
-    /// Create an audit logger that can be used to capture and push
-    /// audit logs to the appropriate logging system over HTTP
-    fn new() -> AuditLogger {
-        // Try to construct the log endpoint URL by joining the base URL
-        // with the ingest path, This can fail if the URL is not valid,
-        // when the base URL is not set or the ingest path is not valid
-        let log_endpoint = CONFIG.parseable.audit_logger.as_ref().and_then(|endpoint| {
-            endpoint
-                .join("/api/v1/ingest")
-                .inspect_err(|err| eprintln!("Couldn't setup audit logger: {err}"))
-                .ok()
-        });
-
-        AuditLogger {
-            log_endpoint,
-            batch: Mutex::new(VecDeque::with_capacity(
-                CONFIG.parseable.audit_batch_size * 10,
-            )),
-        }
-    }
-
-    /// Flushes audit logs to the remote logging system
-    async fn flush(&self) {
-        let Some(endpoint) = self.log_endpoint.as_ref() else {
-            return;
-        };
-
-        let mut logs_to_send = VecDeque::with_capacity(CONFIG.parseable.audit_batch_size * 10);
-        {
-            let mut batch = self.batch.lock().await;
-            if batch.is_empty() {
-                return;
-            }
-
-            std::mem::swap(&mut *batch, &mut logs_to_send)
-        }
-
-        let mut req = HTTP_CLIENT
-            .post(endpoint.as_str())
-            .json(&logs_to_send)
-            .header("x-p-stream", "audit_log");
-
-        // Use basic auth if credentials are configured
-        if let Some(username) = CONFIG.parseable.audit_username.as_ref() {
-            req = req.basic_auth(username, CONFIG.parseable.audit_password.as_ref())
-        }
-
-        // Send batched logs to the audit logging backend
-        let err = match req.send().await {
-            Ok(r) => match r.error_for_status() {
-                Err(e) => e,
-                _ => return,
-            },
-            Err(e) => e,
-        };
-
-        // Attempt to recover the logs on facing an error
-        error!("Failed to send batch of audit logs: {}", err);
-        let mut batch = self.batch.lock().await;
-        batch.extend(logs_to_send);
-    }
-
-    /// Inserts an audit log into the batch, and flushes the batch if it exceeds the configured batch size
-    async fn insert(&self, log: AuditLog) {
-        let mut batch = self.batch.lock().await;
-        batch.push_back(log);
-
-        // Flush if batch size exceeds threshold
-        if batch.len() >= CONFIG.parseable.audit_batch_size {
-            drop(batch); // ensure the lock is released
-            tokio::spawn(AUDIT_LOGGER.flush());
-        }
-    }
-
-    /// Spawns a background task for periodic flushing of audit logs, if configured
-    pub async fn spawn_batcher() {
-        if AUDIT_LOGGER.log_endpoint.is_none() {
-            return;
-        }
-        tokio::spawn(async {
-            let mut interval = interval(CONFIG.parseable.audit_flush_interval);
-            loop {
-                interval.tick().await;
-                AUDIT_LOGGER.flush().await;
-            }
-        });
-    }
-}
-
-// Represents the version of the audit log format
-#[non_exhaustive]
-#[repr(u8)]
-#[derive(Debug, Clone, Copy, Serialize, Default)]
-pub enum AuditLogVersion {
-    // NOTE: default should be latest version
-    #[default]
-    V1 = 1,
-}
-
-#[derive(Serialize, Default)]
-pub struct AuditDetails {
-    pub version: AuditLogVersion,
-    pub id: Ulid,
-    pub generated_at: DateTime<Utc>,
-}
-
-#[derive(Serialize, Default)]
-pub struct ServerDetails {
-    pub version: String,
-    pub deployment_id: Ulid,
-}
-
-// Contains information about the actor (user) who performed the action
-#[derive(Serialize, Default)]
-pub struct ActorDetails {
-    pub remote_host: String,
-    pub user_agent: String,
-    pub username: String,
-    pub authorization_method: String,
-}
-
-// Contains details about the HTTP request that was made
-#[derive(Serialize, Default)]
-pub struct RequestDetails {
-    pub stream: String,
-    pub start_time: DateTime<Utc>,
-    pub end_time: DateTime<Utc>,
-    pub method: String,
-    pub path: String,
-    pub protocol: String,
-    pub headers: HashMap<String, String>,
-}
-
-/// Contains information about the response sent back to the client
-#[derive(Default, Serialize)]
-pub struct ResponseDetails {
-    pub status_code: u16,
-    pub error: Option<String>,
-}
-
-/// The main audit log structure that combines all audit information
-#[derive(Serialize)]
-pub struct AuditLog {
-    pub audit: AuditDetails,
-    pub parseable_server: ServerDetails,
-    pub actor: ActorDetails,
-    pub request: RequestDetails,
-    pub response: ResponseDetails,
-}
-
-/// Builder pattern implementation for constructing audit logs
-pub struct AuditLogBuilder {
-    // Used to ensure that log is only constructed if the logger is enabled
-    enabled: bool,
-    inner: AuditLog,
-}
-
-impl Default for AuditLogBuilder {
-    fn default() -> Self {
-        AuditLogBuilder {
-            enabled: AUDIT_LOGGER.log_endpoint.is_some(),
-            inner: AuditLog {
-                audit: AuditDetails {
-                    version: AuditLogVersion::V1,
-                    id: Ulid::new(),
-                    ..Default::default()
-                },
-                parseable_server: ServerDetails {
-                    version: current().released_version.to_string(),
-                    deployment_id: StorageMetadata::global().deployment_id,
-                },
-                request: RequestDetails {
-                    start_time: Utc::now(),
-                    ..Default::default()
-                },
-                actor: ActorDetails::default(),
-                response: ResponseDetails::default(),
-            },
-        }
-    }
-}
-
-impl AuditLogBuilder {
-    /// Sets the remote host for the audit log
-    pub fn with_host(mut self, host: impl Into<String>) -> Self {
-        if self.enabled {
-            self.inner.actor.remote_host = host.into();
-        }
-        self
-    }
-
-    /// Sets the username for the audit log
-    pub fn with_username(mut self, username: impl Into<String>) -> Self {
-        if self.enabled {
-            self.inner.actor.username = username.into();
-        }
-        self
-    }
-
-    /// Sets the user agent for the audit log
-    pub fn with_user_agent(mut self, user_agent: impl Into<String>) -> Self {
-        if self.enabled {
-            self.inner.actor.user_agent = user_agent.into();
-        }
-        self
-    }
-
-    /// Sets the authorization method for the audit log
-    pub fn with_auth_method(mut self, auth_method: impl Into<String>) -> Self {
-        if self.enabled {
-            self.inner.actor.authorization_method = auth_method.into();
-        }
-        self
-    }
-
-    /// Sets the stream for the request details
-    pub fn with_stream(mut self, stream: impl Into<String>) -> Self {
-        if self.enabled {
-            self.inner.request.stream = stream.into();
-        }
-        self
-    }
-
-    /// Sets the request method details
-    pub fn with_method(mut self, method: impl Into<String>) -> Self {
-        if self.enabled {
-            self.inner.request.method = method.into();
-        }
-        self
-    }
-
-    /// Sets the request path
-    pub fn with_path(mut self, path: impl Into<String>) -> Self {
-        if self.enabled {
-            self.inner.request.path = path.into();
-        }
-        self
-    }
-
-    /// Sets the request protocol
-    pub fn with_protocol(mut self, protocol: impl Into<String>) -> Self {
-        if self.enabled {
-            self.inner.request.protocol = protocol.into();
-        }
-        self
-    }
-
-    /// Sets the request headers
-    pub fn with_headers(mut self, headers: impl IntoIterator<Item = (String, String)>) -> Self {
-        if self.enabled {
-            self.inner.request.headers = headers.into_iter().collect();
-        }
-        self
-    }
-
-    /// Sets the response status code
-    pub fn with_status(mut self, status_code: u16) -> Self {
-        if self.enabled {
-            self.inner.response.status_code = status_code;
-        }
-        self
-    }
-
-    /// Sets the response error if any
-    pub fn with_error(mut self, err: impl Display) -> Self {
-        if self.enabled {
-            let error = err.to_string();
-            if !error.is_empty() {
-                self.inner.response.error = Some(error);
-            }
-        }
-        self
-    }
-
-    /// Sends the audit log to the logging server if configured
-    pub async fn send(self) {
-        // ensures that we don't progress if logger is not enabled
-        if !self.enabled {
-            return;
-        }
-
-        // build the audit log
-        let AuditLogBuilder {
-            inner: mut audit_log,
-            ..
-        } = self;
-
-        let now = Utc::now();
-        audit_log.audit.generated_at = now;
-        audit_log.request.end_time = now;
-
-        AUDIT_LOGGER.insert(audit_log).await
-    }
-}
